Split address transformation for easier reading
